+++ data/examples/declaration/type/parens-comments-out.hs
@@ -1,0 +1,4 @@
+type Test =
+  ( -- a
+    x
+  )